+++ backend/routes/index.js
@@ -1,38 +1,16 @@
-const express = require("express");
-const serverResponses = require("../utils/helpers/responses");
-const messages = require("../config/messages");
-const { Todo } = require("../models/todos/todo");
+const fs = require('fs');
+const path = require('path');
 
-const routes = (app) => {
-  const router = express.Router();
-
-  router.post("/todos", (req, res) => {
-    const todo = new Todo({
-      text: req.body.text,
+module.exports = (app) => {
+  // Lire tous les fichiers du dossier actuel
+  fs.readdirSync(__dirname)
+    .filter((file) => {
+      // Filtre pour obtenir uniquement les fichiers .js, à l'exception de index.js
+      return file.endsWith('.js') && file !== 'index.js';
+    })
+    .forEach((file) => {
+      // Importer chaque fichier de route et l'exécuter avec l'instance app
+      const route = require(path.join(__dirname, file));
+      route(app);
     });
-
-    todo
-      .save()
-      .then((result) => {
-        serverResponses.sendSuccess(res, messages.SUCCESSFUL, result);
-      })
-      .catch((e) => {
-        serverResponses.sendError(res, messages.BAD_REQUEST, e);
-      });
-  });
-
-  router.get("/", (req, res) => {
-    Todo.find({}, { __v: 0 })
-      .then((todos) => {
-        serverResponses.sendSuccess(res, messages.SUCCESSFUL, todos);
-      })
-      .catch((e) => {
-        serverResponses.sendError(res, messages.BAD_REQUEST, e);
-      });
-  });
-
-  //it's a prefix before api it is useful when you have many modules and you want to
-  //differentiate b/w each module you can use this technique
-  app.use("/api", router);
 };
-module.exports = routes;
